pyproject.toml
[project]
name = "semantic-cache"
version = "0.1.0"
description = "Semantic cache for image generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "anthropic>=0.68.0",
    "diffusers>=0.30.0",
    "fastapi>=0.104.0",
    "numpy>=1.24.0",
    "openai>=1.108.0",
    "pillow>=10.0.0",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.24.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "pydantic>=2.5.0",
    "pip>=25.2",
]
